Clarify pipeline output accounting
The number for single frame only may not be relied upon as
the implementations change, but useful for the time being.

Using the existence of initial_pvi as the ProcessCcd success
is confusing because calibrateImage can fail after generating
initial_pvi.

Using the last product of calibrateImage is not ideal either,
but a little closer.