Clarify pipeline output accounting and improve wording
The number for single-frame-only processing as calculated here
can become wrong as the implementations change, but useful for
the time being.

Using the existence of initial_pvi as the ProcessCcd success
is confusing because calibrateImage can fail after generating
initial_pvi.

Using the last product of calibrateImage is not ideal either,
but a little closer.

Send message only if there are on-sky exposures.
can_see_sky started to be a useable criteria starting w_2024_33.
Using can_see_sky or science_program typically give the same results.